Split Name::Synonymy tests out of name_test.rb (#4708)
Moves the 6 tests exercising Name::Synonymy (synonym_ids,
other_approved_synonyms, best_preferred_synonym, other_author_ids/
other_authors, clear_synonym, more_popular) into
test/models/name/synonymy_test.rb, matching app/models/name/synonymy.rb.

test_homonyms is misleadingly named - its body exercises
other_author_ids/other_authors (Synonymy), not Name::Taxonomy#homonyms -
moved with the rest of this section, not left behind.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
